requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
django-crispy-forms==2.3
crispy-bootstrap5==0.7
Pillow==10.2.0
qrcode==7.4.2
openpyxl==3.1.2
reportlab==4.0.7
django-extensions==3.2.3
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework
